Clean up redundant code constructs
Results of the following IntelliJ inspections:
 * Redundant 'isInstance' or 'cast' call
 * Redundant array creation
 * Redundant Collection operation
 * Redundant String operation
 * Redundant type cast

The only manual change was to remove a suppression and comment about a
redundant cast which doesn't appear to be an issue in the current
version of JDK 8

+++ java/com/google/gerrit/metrics/dropwizard/MetricJson.java
@@ -137,7 +137,7 @@ class MetricJson {
       p99_9 = s.get999thPercentile();
 
       min = (double) s.getMin();
-      avg = (double) s.getMean();
+      avg = s.getMean();
       max = (double) s.getMax();
       sum = s.getMean() * m.getCount();
       std_dev = s.getStdDev();
